refactor: Config 기능 단위로 분리
