+++ src/main/java/woowacourse/config/ArgumentResolverConfig.java
@@ -1,0 +1,19 @@
+package woowacourse.config;
+
+import java.util.List;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import woowacourse.auth.support.JwtTokenProvider;
+import woowacourse.config.argumentresolver.AuthenticationPrincipalArgumentResolver;
+import woowacourse.config.interceptor.LoginInterceptor;
+
+@Configuration
+public class ArgumentResolverConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new AuthenticationPrincipalArgumentResolver());
+    }
+}
