+++ src/main/java/woowacourse/config/InterceptorConfig.java
@@ -6,20 +6,15 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import woowacourse.auth.support.JwtTokenProvider;
-import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
-import woowacourse.auth.ui.LoginInterceptor;
+import woowacourse.config.argumentresolver.AuthenticationPrincipalArgumentResolver;
+import woowacourse.config.interceptor.LoginInterceptor;
 
 @Configuration
-public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
+public class InterceptorConfig implements WebMvcConfigurer {
     private final JwtTokenProvider jwtTokenProvider;
 
-    public AuthenticationPrincipalConfig(JwtTokenProvider jwtTokenProvider) {
+    public InterceptorConfig(JwtTokenProvider jwtTokenProvider) {
         this.jwtTokenProvider = jwtTokenProvider;
-    }
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new AuthenticationPrincipalArgumentResolver());
     }
 
     @Override
