+++ src/main/java/woowacourse/config/argumentresolver/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,6 @@
-package woowacourse.auth.ui;
+package woowacourse.config.argumentresolver;
+
+import static woowacourse.auth.support.TokenConst.PAYLOAD;
 
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
@@ -8,7 +10,6 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import woowacourse.auth.dto.LoginCustomer;
 import woowacourse.auth.support.AuthenticationPrincipal;
-import woowacourse.auth.support.TokenConst;
 
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
 
@@ -21,7 +22,7 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer,
                                   NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-        String payload = (String) request.getAttribute(TokenConst.PAYLOAD);
+        String payload = (String) request.getAttribute(PAYLOAD);
         return new LoginCustomer(payload);
     }
 }
